vcms/lajornadavideos/requirements.txt
Django
django-compressor
pysolr
pytz
django-libsass
django-rq
django-htmlmin
django-cacheback[rq]